refactor: reorganize proto files and update imports
- Move proto files from internal/authpb to api/authpb package
- Update go_package option in auth.proto
- Regenerate Go code with new package paths
- Update all Go files to use api/authpb instead of internal/authpb
- Regenerate mocks with new import paths
- Add comprehensive Makefile with proto generation, mock generation, testing, and building targets
- Ensure all packages compile with new proto package structure
