+++ internal/server/auth_server.go
@@ -3,7 +3,7 @@ package server
 import (
 	"context"
 
-	"github.com/Koshsky/subs-service/auth-service/internal/authpb"
+	"github.com/Koshsky/subs-service/auth-service/api/authpb"
 	"github.com/Koshsky/subs-service/auth-service/internal/contracts"
 )
